Vela-Android: porting bttool over socket(AF_INET) (1/2)
VELAPLATFO-26432

dependson:4036148

1. Enable Socket(AF_INET) first on Android
2. Fix the improper definition of Struct over Socekt between 64 bits and 32 bits system.
3. Modify the bttool to let it work similar to Vela

+++ service/ipc/socket/src/bt_socket_server.c
@@ -190,7 +190,7 @@ static int bt_socket_server_receive(service_poll_t* poll, int fd, void* userdata
         bt_socket_server_l2cap_process(poll, fd, ins, &packet);
 #endif
     } else {
-        BT_LOGE("%s, Unhandled message:%d", __func__, packet.code);
+        BT_LOGE("%s, Unhandled message:%d", __func__, (int)packet.code);
         assert(0);
         return BT_STATUS_PARM_INVALID;
     }
